35.1:更新SDK,完善调试所需要的信息;35.2:优化规范代码，删除无用代码，增加log和注释
