+++ AndroidManifest.xml
@@ -1,18 +1,17 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="org.zywx.wbpalmstar.widgetone.uexweixin"
+    package="net.sourceforge.simcpux"
+    android:installLocation="internalOnly"
     android:versionCode="100"
-    android:versionName="3.0.0" 
-    android:installLocation="internalOnly">
+    android:versionName="3.0.0" >
 
     <application
         android:name="org.zywx.wbpalmstar.widgetone.WidgetOneApplication"
         android:allowClearUserData="false"
-        android:debuggable="true"
-        android:icon="@drawable/icon"
-        android:label="@string/app_name" 
         android:hardwareAccelerated="true"
-        android:vmSafeMode="true">
+        android:icon="@drawable/icon"
+        android:label="@string/app_name"
+        android:vmSafeMode="true" >
         <activity
             android:name="org.zywx.wbpalmstar.engine.EBrowserActivity"
             android:alwaysRetainTaskState="true"
@@ -21,74 +20,81 @@
             android:screenOrientation="portrait"
             android:theme="@style/browser_main_theme"
             android:windowSoftInputMode="stateHidden|adjustPan" >
-			<intent-filter>
+            <intent-filter>
                 <action android:name="android.intent.action.VIEW" />
+
                 <category android:name="android.intent.category.DEFAULT" />
                 <category android:name="android.intent.category.BROWSABLE" />
-                <data android:scheme="appcanscheme"/>
+
+                <data android:scheme="appcanscheme" />
             </intent-filter>
         </activity>
-         <activity
+        <activity
             android:name="org.zywx.wbpalmstar.platform.mam.PolicyActivity"
-            android:configChanges="orientation|keyboardHidden|navigation" android:windowSoftInputMode="adjustPan"/>
-         <activity
+            android:configChanges="orientation|keyboardHidden|navigation"
+            android:windowSoftInputMode="adjustPan" />
+        <activity
             android:name="org.zywx.wbpalmstar.platform.mam.PolicyInfoActivity"
-            android:configChanges="orientation|keyboardHidden|navigation" android:windowSoftInputMode="adjustPan"/>
-        <activity android:name="org.zywx.wbpalmstar.engine.TempActivity"
+            android:configChanges="orientation|keyboardHidden|navigation"
+            android:windowSoftInputMode="adjustPan" />
+        <activity
+            android:name="org.zywx.wbpalmstar.engine.TempActivity"
             android:launchMode="standard"
             android:screenOrientation="portrait"
-            android:theme="@style/browser_loading_theme"/>
-        <activity android:name="org.zywx.wbpalmstar.engine.LoadingActivity"
+            android:theme="@style/browser_loading_theme" />
+        <activity
+            android:name="org.zywx.wbpalmstar.engine.LoadingActivity"
             android:configChanges="keyboardHidden|orientation"
-            android:theme="@style/browser_loading_theme"
-            android:screenOrientation="portrait"
             android:launchMode="standard"
-            >
+            android:screenOrientation="portrait"
+            android:theme="@style/browser_loading_theme" >
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <receiver
-            android:name="org.zywx.wbpalmstar.platform.push.PushRecieveMsgReceiver">
+
+        <receiver android:name="org.zywx.wbpalmstar.platform.push.PushRecieveMsgReceiver" >
             <intent-filter>
                 <action android:name="org.zywx.push.receive" />
             </intent-filter>
         </receiver>
-        <!--此处注释掉是为了本地调试方便，若不注释掉，分享完成返回的时候会报错，但不影响大众版在线打包-->
-        <!--<activity
+
+        <!-- uexWeiXin -->
+        <activity
             android:name=".wxapi.WXEntryActivity"
             android:exported="true"
-            android:theme="@style/plugin_uexweixin_Transparent"
-            android:launchMode="singleTop" >
+            android:launchMode="singleTop"
+            android:theme="@style/plugin_uexweixin_Transparent" >
         </activity>
         <activity
             android:name=".wxapi.WXPayEntryActivity"
             android:exported="true"
-             android:theme="@style/plugin_uexweixin_Transparent"
-            android:launchMode="singleTop" >
-        </activity>-->
+            android:launchMode="singleTop"
+            android:theme="@style/plugin_uexweixin_Transparent" >
+        </activity>
     </application>
 
     <uses-sdk android:minSdkVersion="4" />
 
     <uses-permission android:name="android.permission.INTERNET" />
     <uses-permission android:name="android.permission.WRITE_EXTERNAL_STORAGE" />
-	<uses-permission android:name="android.permission.READ_PHONE_STATE" />
-	<uses-permission android:name="android.permission.ACCESS_WIFI_STATE" />
+    <uses-permission android:name="android.permission.READ_PHONE_STATE" />
+    <uses-permission android:name="android.permission.ACCESS_WIFI_STATE" />
     <uses-permission android:name="android.permission.CHANGE_WIFI_STATE" />
-    <uses-permission android:name="android.permission.ACCESS_NETWORK_STATE"/>
+    <uses-permission android:name="android.permission.ACCESS_NETWORK_STATE" />
     <uses-permission android:name="android.permission.READ_LOGS" />
-    <uses-permission android:name="android.permission.RECEIVE_BOOT_COMPLETED"/>
+    <uses-permission android:name="android.permission.RECEIVE_BOOT_COMPLETED" />
     <uses-permission android:name="android.permission.WAKE_LOCK" />
     <uses-permission android:name="android.permission.VIBRATE" />
+
     <supports-screens
         android:anyDensity="true"
         android:largeScreens="true"
-        android:xlargeScreens="true"
         android:normalScreens="true"
         android:resizeable="true"
-        android:smallScreens="true" />
+        android:smallScreens="true"
+        android:xlargeScreens="true" />
 
 </manifest>
