+++ src/net/sourceforge/simcpux/wxapi/WXEntryActivity.java
