+++ src/net/sourceforge/simcpux/wxapi/WXPayEntryActivity.java
