--- src/org/zywx/wbpalmstar/plugin/uexweixin/CheckModel.java
@@ -1,6 +1,0 @@
-package org.zywx.wbpalmstar.plugin.uexweixin;
-
-public class CheckModel {
-	public String errcode;
-	public String errmsg;
-}
