--- src/org/zywx/wbpalmstar/plugin/uexweixin/MD5.java
@@ -1,29 +1,0 @@
-package org.zywx.wbpalmstar.plugin.uexweixin;
-
-import java.security.MessageDigest;
-
-public class MD5 {
-
-	private MD5() {}
-	
-	public final static String getMessageDigest(byte[] buffer) {
-		char hexDigits[] = { '0', '1', '2', '3', '4', '5', '6', '7', '8', '9', 'a', 'b', 'c', 'd', 'e', 'f' };
-		try {
-			MessageDigest mdTemp = MessageDigest.getInstance("MD5");
-			mdTemp.update(buffer);
-			byte[] md = mdTemp.digest();
-			int j = md.length;
-			char str[] = new char[j * 2];
-			int k = 0;
-			for (int i = 0; i < j; i++) {
-				byte byte0 = md[i];
-				str[k++] = hexDigits[byte0 >>> 4 & 0xf];
-				str[k++] = hexDigits[byte0 & 0xf];
-			}
-			return new String(str);
-		} catch (Exception e) {
-			e.printStackTrace();
-			return null;
-		}
-	}
-}
