+++ src/org/zywx/wbpalmstar/plugin/uexweixin/utils/NetworkUtil.java
@@ -1,11 +1,8 @@
-package org.zywx.wbpalmstar.plugin.uexweixin;
+package org.zywx.wbpalmstar.plugin.uexweixin.utils;
 
 import android.app.AlertDialog;
 import android.content.Context;
 import android.content.DialogInterface;
-import android.content.SharedPreferences;
-import android.graphics.Bitmap;
-import android.graphics.Bitmap.CompressFormat;
 import android.util.Log;
 
 import org.apache.http.HttpResponse;
@@ -36,7 +33,6 @@ import java.net.UnknownHostException;
 import java.security.KeyManagementException;
 import java.security.KeyStore;
 import java.security.KeyStoreException;
-import java.security.MessageDigest;
 import java.security.NoSuchAlgorithmException;
 import java.security.UnrecoverableKeyException;
 import java.security.cert.X509Certificate;
@@ -45,10 +41,10 @@ import javax.net.ssl.SSLContext;
 import javax.net.ssl.TrustManager;
 import javax.net.ssl.X509TrustManager;
 
-public class Utils {
-	
+public class NetworkUtil {
+
 	private static final String TAG = "Utils";
-	
+
 	public static byte[] httpGet(final String url) {
 		if (url == null || url.length() == 0) {
 			Log.e(TAG, "httpGet, url is null");
@@ -73,22 +69,22 @@ public class Utils {
 			return null;
 		}
 	}
-	
+
 	public static byte[] httpPost(String url, String entity) {
 		if (url == null || url.length() == 0) {
 			Log.e(TAG, "httpPost, url is null");
 			return null;
 		}
-		
+
 		HttpClient httpClient = getNewHttpClient();
-		
+
 		HttpPost httpPost = new HttpPost(url);
-		
+
 		try {
 			httpPost.setEntity(new StringEntity(entity));
 			httpPost.setHeader("Accept", "application/json");
 			httpPost.setHeader("Content-type", "application/json");
-			
+
 			HttpResponse resp = httpClient.execute(httpPost);
 			if (resp.getStatusLine().getStatusCode() != HttpStatus.SC_OK) {
 				Log.e(TAG, "httpGet fail, status code = " + resp.getStatusLine().getStatusCode());
@@ -103,66 +99,66 @@ public class Utils {
 		}
 	}
 
-	private static HttpClient getNewHttpClient() { 
-		   try { 
-		       KeyStore trustStore = KeyStore.getInstance(KeyStore.getDefaultType()); 
-		       trustStore.load(null, null); 
+	private static HttpClient getNewHttpClient() {
+		try {
+			KeyStore trustStore = KeyStore.getInstance(KeyStore.getDefaultType());
+			trustStore.load(null, null);
 
-		       SSLSocketFactory sf = new SSLSocketFactoryEx(trustStore); 
-		       sf.setHostnameVerifier(SSLSocketFactory.ALLOW_ALL_HOSTNAME_VERIFIER); 
+			SSLSocketFactory sf = new SSLSocketFactoryEx(trustStore);
+			sf.setHostnameVerifier(SSLSocketFactory.ALLOW_ALL_HOSTNAME_VERIFIER);
 
-		       HttpParams params = new BasicHttpParams(); 
-		       HttpProtocolParams.setVersion(params, HttpVersion.HTTP_1_1); 
-		       HttpProtocolParams.setContentCharset(params, HTTP.UTF_8); 
+			HttpParams params = new BasicHttpParams();
+			HttpProtocolParams.setVersion(params, HttpVersion.HTTP_1_1);
+			HttpProtocolParams.setContentCharset(params, HTTP.UTF_8);
 
-		       SchemeRegistry registry = new SchemeRegistry(); 
-		       registry.register(new Scheme("http", PlainSocketFactory.getSocketFactory(), 80)); 
-		       registry.register(new Scheme("https", sf, 443)); 
+			SchemeRegistry registry = new SchemeRegistry();
+			registry.register(new Scheme("http", PlainSocketFactory.getSocketFactory(), 80));
+			registry.register(new Scheme("https", sf, 443));
 
-		       ClientConnectionManager ccm = new ThreadSafeClientConnManager(params, registry); 
+			ClientConnectionManager ccm = new ThreadSafeClientConnManager(params, registry);
 
-		       return new DefaultHttpClient(ccm, params); 
-		   } catch (Exception e) { 
-		       return new DefaultHttpClient(); 
-		   } 
+			return new DefaultHttpClient(ccm, params);
+		} catch (Exception e) {
+			return new DefaultHttpClient();
 		}
-	
-	private static class SSLSocketFactoryEx extends SSLSocketFactory {      
-	      
-	    SSLContext sslContext = SSLContext.getInstance("TLS");      
-	      
-	    public SSLSocketFactoryEx(KeyStore truststore) throws NoSuchAlgorithmException, KeyManagementException, KeyStoreException, UnrecoverableKeyException {      
-	        super(truststore);      
-	      
-	        TrustManager tm = new X509TrustManager() {      
-	      
-	            public X509Certificate[] getAcceptedIssuers() {      
-	                return null;      
-	            }      
-	      
+	}
+
+	private static class SSLSocketFactoryEx extends SSLSocketFactory {
+
+		SSLContext sslContext = SSLContext.getInstance("TLS");
+
+		public SSLSocketFactoryEx(KeyStore truststore) throws NoSuchAlgorithmException, KeyManagementException, KeyStoreException, UnrecoverableKeyException {
+			super(truststore);
+
+			TrustManager tm = new X509TrustManager() {
+
+				public X509Certificate[] getAcceptedIssuers() {
+					return null;
+				}
+
 				@Override
 				public void checkClientTrusted(X509Certificate[] chain, String authType) throws java.security.cert.CertificateException {
 				}
 
 				@Override
-				public void checkServerTrusted(X509Certificate[] chain,	String authType) throws java.security.cert.CertificateException {
-				}  
-	        };      
-	      
-	        sslContext.init(null, new TrustManager[] { tm }, null);      
-	    }      
-	      
+				public void checkServerTrusted(X509Certificate[] chain, String authType) throws java.security.cert.CertificateException {
+				}
+			};
+
+			sslContext.init(null, new TrustManager[] { tm }, null);
+		}
+
 		@Override
 		public Socket createSocket(Socket socket, String host, int port, boolean autoClose) throws IOException, UnknownHostException {
-			return sslContext.getSocketFactory().createSocket(socket, host,	port, autoClose);
+			return sslContext.getSocketFactory().createSocket(socket, host, port, autoClose);
 		}
 
 		@Override
 		public Socket createSocket() throws IOException {
 			return sslContext.getSocketFactory().createSocket();
-		} 
-	}  
-	
+		}
+	}
+
 	public static byte[] streamToByteArray(InputStream in) {
 		if (in == null) {
 			return null;
@@ -183,75 +179,14 @@ public class Utils {
 		return byteArray;
 	}
 
-	public static byte[] bmpToByteArray(Bitmap bmp, boolean needRecycle) {
-		if(bmp==null){
-			return null;
-		}
-		ByteArrayOutputStream baos = new ByteArrayOutputStream();
-		bmp.compress(CompressFormat.PNG, 100, baos);
-		if (needRecycle) {
-			bmp.recycle();
-		}
-
-		byte[] result = baos.toByteArray();
-		try {
-			baos.close();
-		} catch (Exception e) {
-			e.printStackTrace();
-		}
-		return result;
-	}
-	
-	public static AlertDialog showAlert(Context ctx, final String title, final String message, final String label, DialogInterface.OnClickListener listener)
-	{
+	public static AlertDialog showAlert(Context ctx, final String title, final String message, final String label, DialogInterface.OnClickListener listener) {
 		AlertDialog.Builder builder = new AlertDialog.Builder(ctx);
 		builder.setTitle(title);
 		builder.setMessage(message);
 		builder.setPositiveButton(label, listener);
-		AlertDialog alert =builder.create();
+		AlertDialog alert = builder.create();
 		alert.show();
 		return alert;
 	}
-	
-	public static String getAppId(Context ctx){
-		SharedPreferences prefs = ctx.getSharedPreferences("appId", Context.MODE_PRIVATE);
-		return prefs.getString("appId", null);
-	}
-	
-	public static void setAppId(Context ctx, String appId)
-	{
-		SharedPreferences prefs = ctx.getSharedPreferences("appId", Context.MODE_PRIVATE);
-		SharedPreferences.Editor editor = prefs.edit();
-		editor.putString("appId", appId);
-		editor.commit();
-	}
-
-	public static String sha1(String str) {
-		if (str == null || str.length() == 0) {
-			return null;
-		}
-		
-		char hexDigits[] = { '0', '1', '2', '3', '4', '5', '6', '7', '8', '9', 'a', 'b', 'c', 'd', 'e', 'f' };
-		
-		try {
-			MessageDigest mdTemp = MessageDigest.getInstance("SHA1");
-			mdTemp.update(str.getBytes());
-			
-			byte[] md = mdTemp.digest();
-			int j = md.length;
-			char buf[] = new char[j * 2];
-			int k = 0;
-			for (int i = 0; i < j; i++) {
-				byte byte0 = md[i];
-				buf[k++] = hexDigits[byte0 >>> 4 & 0xf];
-				buf[k++] = hexDigits[byte0 & 0xf];
-			}
-			return new String(buf);
-		} catch (Exception e) {
-			e.printStackTrace();
-			return null;
-		}
-	}
-
 
 }
